backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.2.2
twilio==9.10.2
google-genai>=1.0.0
supabase==2.28.0
APScheduler==3.11.2
pydantic==2.12.5
